m2dsia_web_app/requirements.txt
sqlalchemy
pymysql
pydantic
fastapi
uvicorn
pyyaml
